+++ plugins/inputs/kinesis_consumer/kinesis_consumer.go
@@ -1,4 +1,4 @@
-package kinesis
+package kinesis_consumer
 
 import (
 	"context"
